+++ FoldEngine/Editor/Transactions/SetComponentFieldTransaction.cs
@@ -17,12 +17,12 @@ namespace FoldEngine.Editor.Transactions {
         public long EntityId;
 
         public override bool Redo(EditorEnvironment target) {
-            target.Scene.Components.Sets[ComponentType].SetFieldValue((int) EntityId, FieldInfo, NewValue);
+            target.Scene.Components.Sets[ComponentType].SetFieldValue(EntityId, FieldInfo, NewValue);
             return OldValue != NewValue;
         }
 
         public override bool Undo(EditorEnvironment target) {
-            target.Scene.Components.Sets[ComponentType].SetFieldValue((int) EntityId, FieldInfo, OldValue);
+            target.Scene.Components.Sets[ComponentType].SetFieldValue(EntityId, FieldInfo, OldValue);
             return OldValue != NewValue;
         }
     }
